+++ source/CMakeLists.txt
@@ -1,14 +1,14 @@
 if(DEFINED TARGET_LIKE_X86_LINUX_NATIVE)
-    add_library( mbed_trace
+    add_library( mbed-trace
         mbed_trace.c
         mbed_trace_ip6tos.c
     )
     add_definitions("-g -O0 -fprofile-arcs -ftest-coverage")
-    target_link_libraries(mbed_trace gcov)
+    target_link_libraries(mbed-trace gcov)
 else()
-    add_library( mbed_trace
+    add_library( mbed-trace
         mbed_trace.c
         mbed_trace_ip6tos.c
     )
-    target_link_libraries(mbed_trace)
+    target_link_libraries(mbed-trace)
 endif()
